+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add favicon to iodide (fixes #826)
 - Adds JWT-based authentication to Iodide's API as an option (fixes #1755)
+- Rebrand the "jsmd" file format as "iomd" (fixes #1634)
 
 # 0.7.0 (2019-05-30)